Rename GradleProgressIndicatorWrapper to GradleProgressIndicatorEventListener and do some polishing
